Reformat code; Split actions and feedbacks into separate modules
