+++ src/config.ts
@@ -1,4 +1,4 @@
-import { SomeCompanionConfigField } from '@companion-module/base'
+import { Regex, SomeCompanionConfigField } from '@companion-module/base'
 
 export interface IPLOCModuleConfig {
 	host?: string
@@ -12,7 +12,7 @@ export function getConfigFields(): SomeCompanionConfigField[] {
 			id: 'info',
 			width: 12,
 			label: 'Information',
-			value: 'This Module has been tested on IPL-OCP 4.0.0',
+			value: 'Tested with ipl-overlay-controls 4.7.0 running on NodeCG 2.1',
 		},
 		{
 			type: 'textinput',
@@ -20,17 +20,16 @@ export function getConfigFields(): SomeCompanionConfigField[] {
 			label: 'Target host',
 			tooltip: 'The host of the NodeCG instance running IPL OCP',
 			width: 6,
-			default: 'localhost',
+			default: '127.0.0.1',
 		},
 		{
-			type: 'number',
+			type: 'textinput',
 			id: 'port',
 			label: 'Port',
 			tooltip: 'The port of the NodeCG instance running IPL OCP',
 			width: 6,
-			default: 9090,
-			min: 1,
-			max: 65535,
+			regex: Regex.NUMBER,
+			default: '9090',
 		},
 	]
 }
