+++ src/helpers/SplatoonData.ts
@@ -1,50 +1,50 @@
 export const stageNameToShortStageName: Record<string, string> = {
-  'Ancho-V Games': 'Ancho',
-  'Arowana Mall': 'Mall',
-  'Blackbelly Skatepark': 'Blackbelly',
-  'Camp Triggerfish': 'Camp',
-  'Goby Arena': 'Goby',
-  'Humpback Pump Track': 'Humpback',
-  'Inkblot Art Academy': 'Inkblot',
-  'Kelp Dome': 'Kelp',
-  MakoMart: 'Mako',
-  'Manta Maria': 'Manta',
-  'Moray Towers': 'Moray',
-  'Musselforge Fitness': 'Musselforge',
-  'New Albacore Hotel': 'Hotel',
-  'Piranha Pit': 'Pit',
-  'Port Mackerel': 'Port',
-  'Shellendorf Institute': 'Shellendorf',
-  'Shifty Station': 'Shifty',
-  'Snapper Canal': 'Snapper',
-  'Starfish Mainstage': 'Starfish',
-  'Sturgeon Shipyard': 'Sturgeon',
-  'The Reef': 'Reef',
-  'Wahoo World': 'Wahoo',
-  'Walleye Warehouse': 'Walleye',
-  'Skipper Pavilion': 'Skipper',
-  'Unknown Stage': '???',
-  'Museum d\'Alfonsino': 'Museum',
-  'Scorch Gorge': 'Scorch',
-  'Eeltail Alley': 'Eeltail',
-  'Hagglefish Market': 'Haggle',
-  'Undertow Spillway': 'Undertow',
-  'Mincemeat Metalworks': 'Mincemeat',
-  'Hammerhead Bridge': 'Hammerhead',
-  'Mahi-Mahi Resort': 'Mahi',
-  'Brinewater Springs': 'Brinewater',
-  'Flounder Heights': 'Flounder',
-  'Um\'ami Ruins': 'Um\'ami',
-  'Barnacle & Dime': 'Barnacle',
-  'Crableg Capital': 'Crableg',
-  'Shipshape Cargo Co.': 'Shipshape'
+	'Ancho-V Games': 'Ancho',
+	'Arowana Mall': 'Mall',
+	'Blackbelly Skatepark': 'Blackbelly',
+	'Camp Triggerfish': 'Camp',
+	'Goby Arena': 'Goby',
+	'Humpback Pump Track': 'Humpback',
+	'Inkblot Art Academy': 'Inkblot',
+	'Kelp Dome': 'Kelp',
+	MakoMart: 'Mako',
+	'Manta Maria': 'Manta',
+	'Moray Towers': 'Moray',
+	'Musselforge Fitness': 'Musselforge',
+	'New Albacore Hotel': 'Hotel',
+	'Piranha Pit': 'Pit',
+	'Port Mackerel': 'Port',
+	'Shellendorf Institute': 'Shellendorf',
+	'Shifty Station': 'Shifty',
+	'Snapper Canal': 'Snapper',
+	'Starfish Mainstage': 'Starfish',
+	'Sturgeon Shipyard': 'Sturgeon',
+	'The Reef': 'Reef',
+	'Wahoo World': 'Wahoo',
+	'Walleye Warehouse': 'Walleye',
+	'Skipper Pavilion': 'Skipper',
+	'Unknown Stage': '???',
+	"Museum d'Alfonsino": 'Museum',
+	'Scorch Gorge': 'Scorch',
+	'Eeltail Alley': 'Eeltail',
+	'Hagglefish Market': 'Haggle',
+	'Undertow Spillway': 'Undertow',
+	'Mincemeat Metalworks': 'Mincemeat',
+	'Hammerhead Bridge': 'Hammerhead',
+	'Mahi-Mahi Resort': 'Mahi',
+	'Brinewater Springs': 'Brinewater',
+	'Flounder Heights': 'Flounder',
+	"Um'ami Ruins": "Um'ami",
+	'Barnacle & Dime': 'Barnacle',
+	'Crableg Capital': 'Crableg',
+	'Shipshape Cargo Co.': 'Shipshape',
 }
 
 export const modeNameToShortModeName: Record<string, string> = {
-  'Clam Blitz': 'CB',
-  'Tower Control': 'TC',
-  Rainmaker: 'RM',
-  'Splat Zones': 'SZ',
-  'Turf War': 'TW',
-  'Unknown Mode': '??',
+	'Clam Blitz': 'CB',
+	'Tower Control': 'TC',
+	Rainmaker: 'RM',
+	'Splat Zones': 'SZ',
+	'Turf War': 'TW',
+	'Unknown Mode': '??',
 }
