+++ src/helpers/StringHelper.ts
@@ -1,3 +1,3 @@
 export function isBlank(value?: string | null): boolean {
-  return value === null || value === undefined || String(value).trim() === ''
+	return value === null || value === undefined || String(value).trim() === ''
 }
