+++ src/types/replicant.ts
@@ -1,11 +1,11 @@
 export type BundleMap = {
-  [name: string]: ReplicantMap
+	[name: string]: ReplicantMap
 }
 
 export type ReplicantMap = Record<string, any>
 
 export interface ReplicantMetadata {
-  revision: number
-  schemaSum?: string
-  schema?: Record<string, unknown>
+	revision: number
+	schemaSum?: string
+	schema?: Record<string, unknown>
 }
