+++ src/types/replicants/obsData.d.ts
@@ -6,10 +6,10 @@
  */
 
 export interface ObsData {
-  enabled: boolean
-  status: 'CONNECTED' | 'CONNECTING' | 'NOT_CONNECTED'
-  scenes?: string[]
-  currentScene?: string
-  gameplayScene?: string
-  intermissionScene?: string
+	enabled: boolean
+	status: 'CONNECTED' | 'CONNECTING' | 'NOT_CONNECTED'
+	scenes?: string[]
+	currentScene?: string
+	gameplayScene?: string
+	intermissionScene?: string
 }
