+++ tsconfig.json
@@ -1,12 +1,12 @@
 {
-  "extends": "@companion-module/tools/tsconfig/node18/recommended",
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules/**", "src/**/*spec.ts", "src/**/__tests__/*", "src/**/__mocks__/*"],
-  "compilerOptions": {
-    "outDir": "./dist",
-    "baseUrl": "./",
-    "paths": {
-      "*": ["./node_modules/*"]
-    }
-  }
+	"extends": "@companion-module/tools/tsconfig/node18/recommended",
+	"include": ["src/**/*.ts"],
+	"exclude": ["node_modules/**", "src/**/*spec.ts", "src/**/__tests__/*", "src/**/__mocks__/*"],
+	"compilerOptions": {
+		"outDir": "./dist",
+		"baseUrl": "./",
+		"paths": {
+			"*": ["./node_modules/*"]
+		}
+	}
 }
